requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
Django==3.1.5
django-appconf==1.0.4
django-bootstrap-v5==1.0.0
django-environ==0.4.5
django-imagekit==4.0.2
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
pilkit==2.0
Pillow==8.1.0
promise==2.3
psycopg2==2.8.6
python-environ==0.4.54
pytz==2020.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
text-unidecode==1.3
